Renamed index.js in router to landing and renamed the indexRouter to landingRouter in app.js
