+++ server/app.js
@@ -5,7 +5,7 @@ const cookieParser = require("cookie-parser");
 const logger = require("morgan");
 const mongoose = require('mongoose');
 
-const indexRouter = require("./routes/index");
+const langingRouter = require("./routes/landing");
 const pingRouter = require("./routes/ping");
 
 const { json, urlencoded } = express;
@@ -34,7 +34,7 @@ app.use(urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(join(__dirname, "public")));
 
-app.use("/", indexRouter);
+app.use("/", langingRouter);
 app.use("/ping", pingRouter);
 
 // catch 404 and forward to error handler
